\more pointers

+++ 0x06-pointers_arrays_strings/3-strcmp.c
@@ -12,7 +12,7 @@ int _strcmp(char *s1, char *s2)
 
 	for (i = 0; s1[i] != '\0' && s2[i] != '\0'; i++)
 	{
-		cmp = (s1[i] - s2[i]);
+		cmp = (int)(s1[i] - s2[i]);
 	}
 
 	return (cmp);
